rename hot-reload to devtools
