+++ packages/desktop/Cargo.toml
@@ -89,16 +89,15 @@ core-foundation = "0.9.3"
 objc = "0.2.7"
 
 [features]
-default = ["tokio_runtime", "wry/objc-exception", "hot-reload", "devtools"]
+default = ["tokio_runtime", "wry/objc-exception", "devtools"]
 tokio_runtime = ["dep:tokio"]
 fullscreen = ["wry/fullscreen"]
 transparent = ["wry/transparent"]
-devtools = ["wry/devtools"]
-hot-reload = ["dep:dioxus-devtools", "dioxus-signals"]
+devtools = ["wry/devtools", "dep:dioxus-devtools", "dioxus-signals"]
 gnu = []
 
 [package.metadata.docs.rs]
-features = ["tokio_runtime", "hot-reload"]
+features = ["tokio_runtime", "devtools"]
 cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
 default-features = false
 targets = [
