+++ packages/desktop/src/ipc.rs
@@ -19,7 +19,7 @@ pub enum UserWindowEvent {
 
     /// Handle a hotreload event, basically telling us to update our templates
     #[cfg(all(
-        feature = "hot-reload",
+        feature = "devtools",
         debug_assertions,
         not(target_os = "android"),
         not(target_os = "ios")
