+++ packages/desktop/src/launch.rs
@@ -41,12 +41,7 @@ pub fn launch_virtual_dom_blocking(virtual_dom: VirtualDom, desktop_config: Conf
                 #[cfg(any(target_os = "windows", target_os = "linux", target_os = "macos"))]
                 UserWindowEvent::MudaMenuEvent(evnt) => app.handle_menu_event(evnt),
 
-                #[cfg(all(
-                    feature = "hot-reload",
-                    debug_assertions,
-                    not(target_os = "android"),
-                    not(target_os = "ios")
-                ))]
+                #[cfg(all(feature = "devtools", debug_assertions))]
                 UserWindowEvent::HotReloadEvent(msg) => app.handle_hot_reload_msg(msg),
 
                 UserWindowEvent::Ipc { id, msg } => match msg.method() {
