+++ packages/liveview/Cargo.toml
@@ -41,10 +41,10 @@ tower = { workspace = true }
 dioxus = { workspace = true }
 
 [features]
-default = ["hot-reload", "multi-thread"]
+default = ["devtools", "multi-thread"]
 axum = ["dep:axum"]
 multi-thread = ["tokio/rt-multi-thread"]
-hot-reload = ["dep:dioxus-devtools"]
+devtools = ["dep:dioxus-devtools"]
 
 [[example]]
 name = "axum"
