+++ packages/rsx/.vscode/settings.json
@@ -1,8 +1,4 @@
 {
     "rust-analyzer.check.workspace": false,
-    // "rust-analyzer.check.extraArgs": [
-    //     "--features",
-    //     "hot_reload"
-    // ],
     "rust-analyzer.cargo.features": "all"
 }
