+++ packages/web/src/lib.rs
@@ -39,8 +39,8 @@ mod document;
 #[cfg(feature = "document")]
 pub use document::WebDocument;
 
-#[cfg(all(feature = "hot_reload", debug_assertions))]
-mod hot_reload;
+#[cfg(all(feature = "devtools", debug_assertions))]
+mod devtools;
 
 mod hydration;
 #[allow(unused)]
@@ -67,8 +67,8 @@ pub async fn run(mut virtual_dom: VirtualDom, web_config: Config) -> ! {
         console_error_panic_hook::set_once();
     }
 
-    #[cfg(all(feature = "hot_reload", debug_assertions))]
-    let mut hotreload_rx = hot_reload::init();
+    #[cfg(all(feature = "devtools", debug_assertions))]
+    let mut hotreload_rx = devtools::init();
 
     let runtime = virtual_dom.runtime();
 
@@ -123,7 +123,7 @@ pub async fn run(mut virtual_dom: VirtualDom, web_config: Config) -> ! {
     loop {
         // if virtual dom has nothing, wait for it to have something before requesting idle time
         // if there is work then this future resolves immediately.
-        #[cfg(all(feature = "hot_reload", debug_assertions))]
+        #[cfg(all(feature = "devtools", debug_assertions))]
         let template;
         #[allow(unused)]
         let mut hydration_work: Option<SuspenseMessage> = None;
@@ -137,15 +137,15 @@ pub async fn run(mut virtual_dom: VirtualDom, web_config: Config) -> ! {
                 .flatten();
             let mut rx_hydration = hydration_receiver_iter.select_next_some();
 
-            #[cfg(all(feature = "hot_reload", debug_assertions))]
+            #[cfg(all(feature = "devtools", debug_assertions))]
             #[allow(unused)]
             {
-                let mut hot_reload_next = hotreload_rx.select_next_some();
+                let mut devtools_next = hotreload_rx.select_next_some();
                 select! {
                     _ = work => {
                         template = None;
                     },
-                    new_template = hot_reload_next => {
+                    new_template = devtools_next => {
                         template = Some(new_template);
                     },
                     hydration_data = rx_hydration => {
@@ -158,7 +158,7 @@ pub async fn run(mut virtual_dom: VirtualDom, web_config: Config) -> ! {
                 }
             }
 
-            #[cfg(not(all(feature = "hot_reload", debug_assertions)))]
+            #[cfg(not(all(feature = "devtools", debug_assertions)))]
             #[allow(unused)]
             {
                 select! {
@@ -173,13 +173,13 @@ pub async fn run(mut virtual_dom: VirtualDom, web_config: Config) -> ! {
             }
         }
 
-        #[cfg(all(feature = "hot_reload", debug_assertions))]
+        #[cfg(all(feature = "devtools", debug_assertions))]
         if let Some(hr_msg) = template {
             // Replace all templates
             dioxus_devtools::apply_changes(&virtual_dom, &hr_msg);
 
             if !hr_msg.assets.is_empty() {
-                crate::hot_reload::invalidate_browser_asset_cache();
+                crate::devtools::invalidate_browser_asset_cache();
             }
         }
